Revert "Merge pull request #2916 from Inist-CNRS/typescript/workers"
This reverts commit 81ff1a36b02bdc2443e8527cef46246e5be451dd, reversing
changes made to d2678b2b80ae008f1e97757bd58b9b4b0284d60c.
