+++ .dockerignore
@@ -12,7 +12,6 @@
 !babel.config.js
 !ecosystem.config.js
 !modclean.config.js
-!modclean.workers.config.js
 !jest.config.js
 !jsconfig.json
 !packages
