--- modclean.workers.config.js
@@ -1,162 +1,0 @@
-module.exports = {
-    $default: 'files',
-
-    directories: {
-        patterns: [
-            '__tests__',
-            '__mocks__',
-            'test',
-            'tests',
-            'powered-test',
-            'docs',
-            'doc',
-            '.idea',
-            '.vscode',
-            'website',
-            'images',
-            'assets',
-            'example',
-            'examples',
-            'coverage',
-            '.nyc_output',
-            '.circleci',
-            '.github',
-        ],
-        ignore: [
-            // ... glob patterns to ignore here
-        ],
-    },
-
-    files: {
-        patterns: [
-            '*.markdown',
-            '*.md',
-            '*.mkd',
-            '*.jst',
-            '*.coffee',
-            '*.tgz',
-            '*.swp',
-            '*.xls?x',
-            '*.ppt?x',
-            '*.rtf',
-            '*.js.gz',
-            '*.json~',
-            '*.txt',
-            '*.sh',
-            '*.gnu',
-            '*.java',
-            '*.jar',
-            '*.bat',
-            '*.gyp',
-            '*.gypi',
-            '*.h',
-            '*.c',
-            '*.cc',
-            '*.cpp',
-            '*.spec.js',
-            '*.spec.ts',
-            '*.prettierc.*',
-            '*fixture*',
-            // ... glob patterns here
-        ],
-        ignore: [
-            // ... glob patterns to ignore here
-        ],
-    },
-
-    modclean: {
-        patterns: [
-            'readme',
-            '.npmignore',
-            'license',
-            'license.md',
-            'licence.md',
-            'license.markdown',
-            'licence.markdown',
-            'license-mit',
-            'history.md',
-            'history.markdown',
-            '.gitattributes',
-            '.gitmodules',
-            '.travis.yml',
-            'binding.gyp',
-            'contributing',
-            'component.json',
-            'composer.json',
-            'makefile',
-            'gemfile',
-            'rakefile',
-            '.coveralls.yml',
-            'example',
-            'changelog',
-            'changes',
-            '.jshintrc',
-            'bower.json',
-            '*appveyor.yml',
-            '*.log',
-            '*.tlog',
-            '*.patch',
-            '*.sln',
-            '*.pdb',
-            '*.vcxproj*',
-            '.gitignore',
-            '.sauce-labs*',
-            '.vimrc*',
-            '.idea',
-            'examples',
-            'samples',
-            'test',
-            'tests',
-            'draft-00',
-            'draft-01',
-            'draft-02',
-            'draft-03',
-            'draft-04',
-            '.eslintrc',
-            '.eslintrc.*',
-            '.jamignore',
-            '.jscsrc',
-            '*.todo',
-            '*.md',
-            '*.markdown',
-            '*.js.map',
-            'contributors',
-            '*.orig',
-            '*.rej',
-            '.zuul.yml',
-            '.editorconfig',
-            '.npmrc',
-            '.jshintignore',
-            '.eslintignore',
-            '.lint',
-            '.lintignore',
-            'cakefile',
-            '.istanbul.yml',
-            'authors',
-            'hyper-schema',
-            'mocha.opts',
-            '.gradle',
-            '.tern-port',
-            '.gitkeep',
-            '.dntrc',
-            '*.watchr',
-            '.jsbeautifyrc',
-            'cname',
-            'screenshots',
-            '.dir-locals.el',
-            'jsl.conf',
-            'jsstyle',
-            'benchmark',
-            'dockerfile',
-            '*.nuspec',
-            '*.csproj',
-            'thumbs.db',
-            '.ds_store',
-            'desktop.ini',
-            'npm-debug.log',
-            'wercker.yml',
-            '.flowconfig',
-        ],
-        ignore: ['validate-npm-package-license', 'spdx-license-ids'],
-    },
-};
