+++ package.json
@@ -12,7 +12,7 @@
     "scripts": {
         "start": "concurrently \"npm:production:*\"",
         "production:api": "tsx --max-http-header-size=32768 src/api",
-        "production:workers": "tsx workers",
+        "production:workers": "node --require @babel/register workers",
         "build": "npx vite build --config ./vite.config.ts",
         "development:api": "tsx watch --max-http-header-size=32768 src/api",
         "development:app": "npx vite --port=8080 --host=0.0.0.0",
@@ -22,7 +22,7 @@
         "lint:test": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
         "clean": "concurrently \"npm:clean:*\"",
         "clean:modules": "modclean -Pr -n ./modclean.config.js:*",
-        "clean:workers": "modclean -Pr -n ./modclean.workers.config.js:* -D workers",
+        "clean:workers": "modclean -Pr -n ./modclean.config.js:* -D workers",
         "clean:packages": "modclean -Pr -n ./modclean.config.js:* -D packages",
         "test:unit": "NODE_OPTIONS='--max_old_space_size=8192' ./node_modules/jest/bin/jest.js --force-exit --colors --maxWorkers=1",
         "test:unit:watch": "NODE_OPTIONS='--max_old_space_size=8192' ./node_modules/jest/bin/jest.js --watchAll --colors",
