+++ workers/formatExporters/jsonallvalue.spec.js
@@ -75,8 +75,7 @@ test.skip('export single resource', (done) => {
                 { localConfig: {}, fields },
             ),
         )
-        // @ts-expect-error TS(7006) FIXME: Parameter 'data' implicitly has an 'any' type.
-        .on('data', data => {
+        .on('data', (data) => {
             if (data) outputString += data;
         })
         .on('end', () => {
