+++ workers/loaders/atom.spec.js
@@ -2,19 +2,16 @@ const ezs = require('@ezs/core');
 const from = require('from');
 
 describe('atom.ini', () => {
-    it('should parse an ATOM XML feed', done => {
-        // @ts-expect-error TS(7034) FIXME: Variable 'res' implicitly has type 'any[]' in some... Remove this comment to see the full error message
+    it('should parse an ATOM XML feed', (done) => {
         const res = [];
         from([
             `<feed><entry><any>value</any><other>thing</other></entry></feed>`,
         ])
             .pipe(ezs('delegate', { file: __dirname + '/atom.ini' }))
-            // @ts-expect-error TS(7006) FIXME: Parameter 'chunk' implicitly has an 'any' type.
-            .on('data', chunk => {
+            .on('data', (chunk) => {
                 res.push(chunk);
             })
             .on('end', () => {
-                // @ts-expect-error TS(2304) FIXME: Cannot find name 'expect'.
                 expect(res).toEqual([ { 'any/$t': 'value', 'other/$t': 'thing' } ]); // eslint-disable-line
                 done();
             });
