+++ workers/loaders/csv-comma.spec.js
@@ -2,17 +2,14 @@ const ezs = require('@ezs/core');
 const from = require('from');
 
 describe('csv-comma.ini', () => {
-    it('should parse a CSV', done => {
-        // @ts-expect-error TS(7034) FIXME: Variable 'res' implicitly has type 'any[]' in some... Remove this comment to see the full error message
+    it('should parse a CSV', (done) => {
         const res = [];
         from([`a,b\n1,2\n`])
             .pipe(ezs('delegate', { file: __dirname + '/csv-comma.ini' }))
-            // @ts-expect-error TS(7006) FIXME: Parameter 'chunk' implicitly has an 'any' type.
-            .on('data', chunk => {
+            .on('data', (chunk) => {
                 res.push(chunk);
             })
             .on('end', () => {
-                // @ts-expect-error TS(2304) FIXME: Cannot find name 'expect'.
                 expect(res).toEqual([{ a: '1', b: '2' }]);
                 done();
             });
