+++ workers/loaders/csv.spec.js
@@ -3,31 +3,25 @@ const from = require('from');
 
 describe('csv.ini', () => {
     it('should parse a comma csv', (done) => {
-        // @ts-expect-error TS(7034) FIXME: Variable 'res' implicitly has type 'any[]' in some... Remove this comment to see the full error message
         const res = [];
         from(['a,b\n1,2\n'])
             .pipe(ezs('delegate', { file: __dirname + '/csv.ini' }))
-            // @ts-expect-error TS(7006) FIXME: Parameter 'chunk' implicitly has an 'any' type.
             .on('data', (chunk) => {
                 res.push(chunk);
             })
             .on('end', () => {
-                // @ts-expect-error TS(2304) FIXME: Cannot find name 'expect'.
                 expect(res).toEqual([{ a: '1', b: '2' }]);
                 done();
             });
     });
     it('should parse a semicolon csv', (done) => {
-        // @ts-expect-error TS(7034) FIXME: Variable 'res' implicitly has type 'any[]' in some... Remove this comment to see the full error message
         const res = [];
         from(['a;b\n1;2\n'])
             .pipe(ezs('delegate', { file: __dirname + '/csv.ini' }))
-            // @ts-expect-error TS(7006) FIXME: Parameter 'chunk' implicitly has an 'any' type.
             .on('data', (chunk) => {
                 res.push(chunk);
             })
             .on('end', () => {
-                // @ts-expect-error TS(2304) FIXME: Cannot find name 'expect'.
                 expect(res).toEqual([{ a: '1', b: '2' }]);
                 done();
             });
