+++ workers/routines/sitemap.spec.js
@@ -10,8 +10,7 @@ test.skip('export an xml', (done) => {
         },
     ])
         .pipe(ezs('delegate', { file: __dirname + '/sitemap.ini' }))
-        // @ts-expect-error TS(7006) FIXME: Parameter 'data' implicitly has an 'any' type.
-        .on('data', data => {
+        .on('data', (data) => {
             if (data) outputString += data;
         })
         .on('end', () => {
